port laravel5's container functions to lumen
